nave_analytics/requirements.txt
-e .
attrs==23.2.0
blinker==1.7.0
boto3==1.34.84
botocore==1.34.84
certifi==2024.2.2
cftime==1.6.3
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
ecdsa==0.19.0
fiona==1.9.6
Flask==3.0.3
Flask-Cors==4.0.0
flask-swagger-ui==4.11.1
geopandas==0.14.3
gunicorn==21.2.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
netCDF4==1.6.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pyasn1==0.6.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-jose==3.3.0
pytz==2024.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
scipy==1.13.0
shapely==2.0.3
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
xarray==2024.3.0
